final frontedn no revisi

+++ src/rack/rack.controller.ts
@@ -11,10 +11,13 @@ import {
 } from '@nestjs/common';
 
 import { RackService } from './rack.service';
+
 import { UpdateRackDto } from './dto/update-rack.dto';
 
 import { JwtAuthGuard } from 'src/auth/guard/jwt-auth.guard';
+
 import { RolesGuard } from 'src/auth/guard/roles.guard';
+
 import { Roles } from 'src/auth/Decorators/roles.decorator';
 
 import { Role, Divisi } from '@prisma/client';
@@ -33,7 +36,6 @@ export class RackController {
   @UseGuards(RolesGuard)
   @Roles(Role.super_admin)
   findAllAdminRacks() {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call
     return this.rackService.findAllAdminRacks();
   }
 
@@ -48,7 +50,10 @@ export class RackController {
   }
 
   @Get('divisi/:divisi')
-  findByDivision(@Param('divisi') divisi: string) {
+  findByDivision(
+    @Param('divisi')
+    divisi: string,
+  ) {
     if (!Object.values(Divisi).includes(divisi as Divisi)) {
       throw new BadRequestException('Divisi tidak valid');
     }
@@ -65,24 +70,43 @@ export class RackController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Req() req: any, @Body() dto: UpdateRackDto) {
+  @UseGuards(RolesGuard)
+  @Roles(Role.admin_rack)
+  update(
+    @Param('id') id: string,
+
+    @Req() req: any,
+
+    @Body()
+    dto: UpdateRackDto,
+  ) {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access
     return this.rackService.updateRack(id, req.user.id, dto);
   }
 
   @Delete(':id')
   @UseGuards(RolesGuard)
-  @Roles(Role.super_admin)
-  delete(@Param('id') id: string) {
-    return this.rackService.deleteRack(id);
+  @Roles(Role.super_admin, Role.admin_rack)
+  delete(@Param('id') id: string, @Req() req: any) {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    if (req.user.role === Role.super_admin) {
+      return this.rackService.deleteRack(id);
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access
+    return this.rackService.deleteRack(id, req.user.id);
   }
 
   @Patch(':id/approve')
+  @UseGuards(RolesGuard)
+  @Roles(Role.super_admin)
   approve(@Param('id') id: string) {
     return this.rackService.approveRack(id);
   }
 
   @Patch(':id/reject')
+  @UseGuards(RolesGuard)
+  @Roles(Role.super_admin)
   reject(@Param('id') id: string) {
     return this.rackService.rejectRack(id);
   }
